+++ wp-content/plugins/cat-game-vote-standalone/includes/class-reports.php
@@ -1410,6 +1410,8 @@ class CatGame_Reports {
         delete_user_meta($user_id, 'catgame_country');
         delete_user_meta($user_id, 'catgame_custom_tags');
         delete_user_meta($user_id, 'catgame_user_tags');
+        delete_user_meta($user_id, 'catgame_terms_accepted');
+        delete_user_meta($user_id, 'catgame_terms_accepted_at');
         delete_user_meta($user_id, 'catgame_avatar_color');
     }